fixed a few issues

+++ client/web/__init__.py
@@ -15,7 +15,7 @@ def get():
     d.serial.get(q)
     d.serial.loop()
     vals = q.get_nowait()
-    return jsonify(time=round(vals[0]*1000),
+    return jsonify(time=int(round(vals[0]*1000)),
             voltage=vals[1], motorAmps=vals[2], chargeAmps=vals[3], kwhs=vals[4])
 
 @app.route('/command')
